// src/app/models/company.model.ts
export class CompanyModel {
  compCode: number;
  compName: string;
  compPhone: string;
  contactPersonName: string;
  contactPersonPhone: string;
  contactPersonEmail: string;
  compAddress: string;
  altContactPerson: string;
  altContactPhone: string;
  altContactEmail: string;
  panTan: string;
  compRegCode: string;
  serviceTax: string;

  mouSigned: number;
  compStatus: string;
  compWebsite: string;

  effectiveDate: Date;
  expirationDate: Date;
  createdAt: Date;
  updatedAt: Date;
}